% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.combine.species.r
\name{comb.species}
\alias{comb.species}
\title{Combine species in vegetation matrix}
\usage{
comb.species(x, sel, newname, refl)
}
\arguments{
\item{x}{(veg) vegetation matrix of class veg}

\item{sel}{(character) vector of species to combine}

\item{newname}{(character) name of the new taxon}

\item{refl}{(character) Name of taxonomic reference list}
}
\description{
Combine species in vegetation matrix
}
