% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elbaue}
\alias{elbaue}
\title{This is an example vegetation dataset to be included in package vegdata. Use tv.veg('elbaue') and tv.site('elbaue') to load.}
\source{
\url{https://www.vegetweb.de}
}
\description{
This is an example vegetation dataset to be included in package vegdata. Use tv.veg('elbaue') and tv.site('elbaue') to load.
}
\author{
Florian Jansen
}
\keyword{data}
