% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lc.0}
\alias{lc.0}
\title{Layer combinations}
\format{
A data frame with rows and 2 variables:
\describe{
  \item{LAYER}{Layer code, i.e. 0:9 for Turboveg}
  \item{COMB}{Combinations. Same integer means, they will be combined}
}
}
\usage{
lc.0
}
\description{
Example tables with layer codes and how they should be combined
 in vegetation analyses. lc.0: do not combine any layers
}
\keyword{datasets}
