% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lc.1}
\alias{lc.1}
\title{Layer combinations}
\format{
A data frame with rows and 3 variables:
\describe{
  \item{LAYER}{Layer code, i.e. 0:9 for Turboveg}
  \item{COMB}{Combinations. Same integer means, they will be combined}
}
}
\usage{
lc.1
}
\description{
datasets with layer codes and how they should be combined
 in vegetation analyses.
}
\keyword{datasets}
