% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lc.all}
\alias{lc.all}
\title{Layer combinations}
\format{
A data frame with rows and 2 variables:
\describe{
  \item{LAYER}{Layer code, i.e. 0:9 for Turboveg}
  \item{COMB}{Combinations. Same integer means, they will be combined}
}
}
\usage{
lc.all
}
\description{
A data.frame to combine all layers in a vegetation plot
}
\keyword{datasets}
