% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.names.R
\name{parse.taxa}
\alias{parse.taxa}
\title{Parse taxon strings into genus part and epitheta}
\usage{
parse.taxa(x, epis)
}
\arguments{
\item{x}{(character) taxon names}

\item{epis}{(character) vector of separators for epithets (like e.g. "subsp.")}
}
\description{
Parse taxon strings into genus part and epitheta
}
