% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.names.R
\name{recode.species}
\alias{recode.species}
\title{Recode species names, lettercodes or ID's}
\usage{
recode.species(
  x,
  names = c("shortletters", "Numbers", "ScientificNames"),
  refl
)
}
\arguments{
\item{x}{vector of species}

\item{names}{one of 7digit shortletter codes, species id's or scientific species names}

\item{refl}{(character) name of taxon reference list}
}
\description{
Recode species names, lettercodes or ID's
}
