% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_vegdata}
\alias{src_vegdata}
\alias{src_eurosl}
\alias{src_germansl}
\title{src - dplyr src objects}
\usage{
src_eurosl(path = db_path("eurosl"), ...)

src_germansl(path = db_path("germansl"), ...)
}
\arguments{
\item{path}{(character) path to SQLite database. by default
we use the function [db_path()] to get the path}

\item{...}{Further args passed on to [DBI::dbConnect()]}
}
\value{
an src object
}
\description{
src - dplyr src objects
}
\examples{
\dontrun{
# src_eurosl()
# src_germansl()
}
}
