% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.reflist.r
\name{tax.refl}
\alias{tax.refl}
\title{Taxon reference list to be used}
\usage{
tax.refl(refl, db, tv_home)
}
\arguments{
\item{refl}{name of taxonomic reference list}

\item{db}{optional: Turboveg database name}

\item{tv_home}{Turboveg installation path}
}
\description{
Taxon reference list to be used
}
