% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxlevels}
\alias{taxlevels}
\title{Taxonomic hierarchy levels}
\format{
A data frame with rows and 4 variables:
\describe{
  \item{rank}{Layer code, i.e. 0:9 for Turboveg}
  \item{level}{Combinations. Same integer means, they will be combined}
  \item{Euro.Med}{taxon level name in http://www.europlusmed.org/ database}
  \item{description}{Explanation of level codes}
}
}
\usage{
taxlevels
}
\description{
hierarchy of taxon levels
}
\keyword{datasets}
