% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxref_download.R
\name{taxref_download}
\alias{taxref_download}
\alias{db_download}
\alias{taxref_download_eurosl}
\alias{taxref_download_germansl}
\title{Download taxonomic databases}
\usage{
taxref_download_eurosl(version = "latest", verbose = TRUE, overwrite = FALSE)

taxref_download_germansl(version = "1.5", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{version}{(character) desired version number of the list}

\item{verbose}{(logical) Print messages. Default: `TRUE`}

\item{overwrite}{(logical) If `TRUE` force an update by overwriting
previously downloaded data. Default: `FALSE`}
}
\value{
(character) path to the downloaded SQL database
}
\description{
Download taxonomic databases
}
\details{
Downloads sql database, cleans up unneeded files, returns path
to sql file
}
\examples{
\dontrun{
# EuroSL
taxref_download_eurosl()
src_eurosl()

# GermanSL
taxref_download_germansl()
taxref_download_germansl(overwrite=TRUE) # overwrite - download again
src_germansl()
}
}
\seealso{
[.my_cache]
}
