% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.names.R
\name{TCS.replace}
\alias{TCS.replace}
\title{Standardise taxon list field names to match the Taxonomic Concept Transfer Schema (TCS)}
\usage{
TCS.replace(x)
}
\arguments{
\item{x}{(character) string of column names used in data.frames storing taxon lists}
}
\description{
Applies Taxonomic Concept Transfer Schema (TCS) to the different name list conventions of different sources
}
\references{
Taxonomic Names and Concepts interest group. 2006. Taxonomic Concept Transfer Schema (TCS), version 1.01. Biodiversity Information Standards (TDWG) http://www.tdwg.org/standards/117
}
\author{
Florian Jansen florian.jansen@uni-rostock.de
}
