% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.biblio.R
\name{tv.bib}
\alias{tv.bib}
\title{Check bibliographic references from Turboveg codes}
\usage{
tv.bib(x = "all", db, dict = tv.dict(db), tv_home, ...)
}
\arguments{
\item{x}{(character) Turboveg reference code(s), e.g. "000001"}

\item{db}{(character) Database name. Needed to select appropriate TV Dictionary folder.}

\item{dict}{(character) Name of Turboveg Dictionary (term lists for header data) if not the default one.}

\item{tv_home}{(character) Turbowin installation path. If not specified function [tv.home()] tries to discover.}

\item{\dots}{additional arguments}
}
\value{
Dataframe of (selected) bibliographic references (when assigned to an object).
}
\description{
Check bibliographic references from Turboveg codes
}
\author{
Florian Jansen \email{florian.jansen@uni-rostock.de}
}
\keyword{Turboveg}
