% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.coverperc.R
\name{tv.coverperc}
\alias{tv.coverperc}
\title{Cover code translation}
\usage{
tv.coverperc(db, obs, RelScale, tv_home, tvscale, ...)
}
\arguments{
\item{db}{the name of the Turboveg database}

\item{obs}{dataframe of observations, containing Cover Codes, coded in tvscale.dbf of Turboveg installation}

\item{RelScale}{Dataframe of CoverScale codes per releve, if empty it is read from the database}

\item{tv_home}{Path to Turboveg installation}

\item{tvscale}{Cover scale}

\item{\dots}{Further options}
}
\value{
data.frame of observations with additional column COVER_PERC
}
\description{
Translate cover code into percentage cover values for Turboveg database observations.
}
\keyword{Turboveg}
