% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.db.R
\name{tv.db}
\alias{tv.db}
\alias{tv.dict}
\title{Read list of available Turboveg2 databases in given Turboveg directory}
\usage{
tv.db(path)

tv.dict(db, tv_home)
}
\arguments{
\item{path}{(character) directory path inside Turboveg/data directory}

\item{db}{(character) name of Turboveg database/directory}

\item{tv_home}{(character) path of Turboveg instalation}
}
\value{
List of databases below specified path
}
\description{
Read list of available Turboveg2 databases in given Turboveg directory

Read list of available Turboveg2 databases in given Turboveg directory
}
\author{
Florian Jansen \email{florian.jansen@uni-rostock.de}
}
\keyword{Turboveg}
