% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.readXML.R
\name{tv.readXML}
\alias{tv.readXML}
\title{Reads a Turboveg XML file}
\usage{
tv.readXML(file)
}
\arguments{
\item{file}{(character) Path name of the Turboveg XML file}
}
\value{
S3 list with elements tvwin, tvadmin, site, and obs
}
\description{
Reads Turboveg XML formatted files species-plot observations and site information into a list
}
\seealso{
\code{\link{tv.veg}}, \code{\link{tv.site}}
}
\author{
Florian Jansen \email{florian.jansen@uni-rostock.de}
}
\keyword{Turboveg}
