% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.reflist.r
\name{tv.refl}
\alias{tv.refl}
\title{Taxon reference list to be used}
\usage{
tv.refl(refl, db, tv_home)
}
\arguments{
\item{refl}{name of reference list}

\item{db}{Turboveg database name}

\item{tv_home}{Turboveg installation path}
}
\description{
Taxon reference list to be used
}
