% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.names.R
\name{TV.replace}
\alias{TV.replace}
\title{Rename data.frame columns to match Turboveg 2 conventions}
\usage{
TV.replace(x)
}
\arguments{
\item{x}{(character) string vector of column names}
}
\description{
Rename data.frame columns to match Turboveg 2 conventions
}
