\name{agg}
\alias{agg}

\title{Show members of specified taxa}

\description{
Show the members of taxonomical levels, e.g. aggregates \code{AGG}.
}

\usage{
agg(member='all', refl, tv_home,...)
}

\arguments{
\item{member}{Species number, lettercode or name}
\item{refl}{Taxonomic reference list to use}
\item{tv_home}{Directory path of Turbowin installation}
\item{...}{additional arguments}
}

\author{Florian Jansen
\email{jansen@uni-greifswald.de}
        }

\keyword{misc}