% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.site.R
\docType{data}
\name{elbaue}
\alias{elbaue}
\title{This is an example vegetation dataset to be included in package vegdata}
\description{
This is an example vegetation dataset to be included in package vegdata
}
\references{
\url{https://www.vegetweb.de}
}
\author{
Florian Jansen \email{florian.jansen@uni-rostock.de}
}
\keyword{data}
