\name{elbaue}
\alias{elbaue}
\alias{taxatest}

\title{Example datasets in Turboveg format}

\description{
Dataset elbaue contains 30 releves from Leyer and Wesche 2007, a selection of releves from (old) floodplain areas of river elbe.

Dataset taxatest is an artificial dataset of three releves to show taxonomic emendation with \code{\link{tv.taxval}}.
}

\seealso{\code{\link{tv.taxval}}, vignette vegdata \code{vignette('vegdata', package='vegdata')} }

\author{Florian Jansen
\email{jansen@uni-greifswald.de}
        }

\keyword{data}