% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.names.R
\name{taxname.removeAuthors}
\alias{taxname.removeAuthors}
\title{Remove name authors from taxon names}
\usage{
taxname.removeAuthors(x, sep)
}
\arguments{
\item{x}{(character) vector of taxon names}

\item{sep}{(character) vector of rank indicators}
}
\description{
Remove name authors from taxon names
}
