\name{tv.obs}
\alias{tv.obs}
\title{Dataframe of plot-species observations directly from Turboveg}

\description{Dataframe of plot-species observations directly from Turboveg.}

\usage{
tv.obs(db, tv_home, ...)
}

\arguments{
\item{db}{Name of your Turboveg database. Directory name containing tvabund.dbf, tvhabita.dbf and tvwin.set. Please specify pathnames below but not above Turbowin/Data.}

\item{tv_home}{Turbowin installation path. Optional, if Turbowin is either on "C:/turbowin" or "C:/Programme/Turbowin"}
\item{\dots}{additional arguments}
}

\value{ Data.frame of species occurences.}

\author{Florian Jansen \email{jansen@uni-greifswald.de}      }

\keyword{misc}
