% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.xml.r
\name{tv.readXML}
\alias{tv.readXML}
\title{Reads a Turboveg XML file and writes a Turboveg dbase database}
\usage{
tv.readXML(file)
}
\arguments{
\item{file}{the filename of an XML file using Turboveg format (Turboveg version xx)}
}
\description{
Reads a Turboveg XML file and writes a Turboveg dbase database
}
\author{Florian Jansen \email{florian.jansen@uni-rostock.de}}

\keyword{misc, survey}
