\name{vegdata-internal}
\alias{tv.home}
\alias{tv.refl}
\alias{sub.abbr}
\alias{vw.con}
\alias{first.word}

\title{Internal vegdata functions}

\description{
  Internal vegdata functions.
}

\details{
  These are not to be called by the user.

  tv.home tries to guess the default tv_home directory (\code{'C:\Turbowin'} or \code{'C:\Programme\Turbowin'} or \code{'O:\Turbowin'} on Windows systems and \code{'~/.wine/drive_c/Turbowin'} on Unix systems.
}

\keyword{internal}