% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegdata-internal.r
\name{vegdata-internal}
\alias{vegdata-internal}
\alias{asc}
\alias{reShape.veg}
\alias{bin2word}
\alias{word2bin}
\title{Internal vegdata functions}
\usage{
asc(char)
}
\description{
Internal vegdata functions.
}
\details{
These are not intended to be called directly by the user.
 tv.home tries to guess the default tv_home directory (\code{'C:\Turbowin'} or \code{'C:\Programme\Turbowin'} or \code{'O:\Turbowin'} on Windows systems and \code{'~/.wine/drive_c/Turbowin'} on Unix systems.
 As dBase is an old DOS format, Umlaute have been stored in Turboveg using the CP437 code table. This has been changed and Turboveg seems to use a country specific code page now. Change options('tv.iconv') if you run into problems
}
\keyword{internal}
