% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegdata-package.R
\docType{package}
\encoding{UTF-8}
\name{vegdata-package}
\alias{vegdata-package}
\alias{vegdata}
\title{vegdata}
\description{
Functions to access data from vegetation databases and evaluate taxon names.
}
\section{Introduction}{

This package provides a set of functions to load data from vegetation databases (at present Turboveg and vegetweb.de).
Taxa can be (semi-)automatically be checked and adapted depending the scientific question. For this a hierachical taxonomic reference list is needed.
}

\section{Taxonomic harmonization}{

Use \code{\link{tv.veg}} to prepare data directly for further analyses. Set option \code{ßlink{taxval}} to \code{TRUE}, if your database is referenced with GermanSL or equivalent taxonomic reference list and you want to realize taxonomic checks and adaptations.
For more details see \code{vignette('vegdata')}.
}

\references{
Jansen, F., Dengler, J (2011) Plant names in vegetation databases - a neglected source of bias,
Journal of vegetation science, 21(6), 1179-1186. @url{http://dx.doi.org/10.1111/j.1654-1103.2010.01209.x}

Jansen, Florian and Dengler, Juergen (2008) GermanSL - eine universelle taxonomische Referenzliste fuer Vegetationsdatenbanken, Tuexenia, 28, 239-253.
}
\author{
Florian Jansen \email{florian.jansen@uni-Rostock.de}
}
\keyword{package}
